This pull request fixes issue #190 by introducing a custom split function that skips empty matches (according to https://www.austingroupbugs.net/view.php?id=1468, this is the correct behaviour). It is used only for FS

Fixes #190