@fioriandrea 2023-10-10
The results of a benchmark I ran manually are the following:

```sh
ls -lh ../data.bench 
-rw-r--r-- 1 af af 353M 10 ott 19.30 ../data.bench

# with custom split
time ./goawk -F '[0123456789!@#$%^&*()_+]+' '{ print $20 }' ../data.bench >/dev/null

real    0m27,727s
user    0m28,734s
sys     0m0,536s

# without custom split
time ./goawk -F '[0123456789!@#$%^&*()_+]+' '{ print $20 }' ../data.bench >/dev/null

real    0m28,043s
user    0m28,245s
sys     0m0,186s
```

I generated the benchmark data with the following script:

```awk
#/bin/awk -f

BEGIN {
    columns = 10000
    rows = 4000
    maxDelimLen = 10
    len = split("0123456789!@#$%^&*()_+", chars, "")
    for (r = 1; r <= rows; r++) {
        for (c = 1; c <= columns; c++) {
            delimLen = int(maxDelimLen * rand()) + 1
            printf("ABCD")
            for (i = 1; i <= delimLen; i++) {
                char = chars[int(len * rand())]
                printf(char)
            }
        }
        printf("ABCD")
        printf("\n")
    }
}
```

Just for curiosity, I tried to run the same benchmark on gawk, and I got the following results:

```sh
time gawk -F '[0123456789!@#$%^&*()_+]+' '{ print $20 }' ../data.bench >/dev/null

real    0m0,228s
user    0m0,183s
sys     0m0,045s
```

I suspect that gawk doesn't split fields if it doesn't need to. In contrast, in goawk all the fields are generated for each line (in this case, split has been called 4000 times, generating a slice containing all fields each time. This is done both with my custom split and with the built-in split).