@benhoyt 2023-10-16
Thanks for this! I made a small tweak to allow reuse of a pre-allocated `p.fields` slice. I'm satisified that at least it's not any/much slower than what we what. Wonder what Gawk is doing there, though! Go's regexp package is fairly slow, but I wouldn't have thought that slow... still, that's a crazy input dataset you have there. :-)